Separação do botão de criar nova rota
