+++ frontend/front_vanFacil/src/screens/Motorista/Inicial.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import { FlatList, StyleSheet, Text, TouchableOpacity } from 'react-native';
-import { Feather } from '@expo/vector-icons';
+import { FlatList, StyleSheet } from 'react-native';
 
 import cores from '../../../assets/cores';
 
 import CardRota from '../Shared/CardRota';
+import BtnNovaRota from '../Shared/Inicial/BtnNovaRota';
 
 import rotas from '../../mocks/rotas';
 
@@ -27,34 +27,3 @@ function ListaRotas() {
       />
    );
 }
-
-function BtnNovaRota(style) {
-   return (
-      <TouchableOpacity style={[style, estilos.botao]}>
-         <Feather name="plus" style={estilos.iconeBotao} />
-      </TouchableOpacity>
-   );
-}
-
-const estilos = StyleSheet.create({
-   lista: {
-      flex: 1,
-      justifyContent: 'center',
-      alignItems: 'center',
-   },
-   botao: {
-      position: 'absolute',
-      bottom: 25,
-      right: 25,
-      alignItems: 'center',
-      justifyContent: 'center',
-      width: 60,
-      height: 60,
-      backgroundColor: cores.azulProfundo,
-      borderRadius: 50,
-   },
-   iconeBotao: {
-      fontSize: 30,
-      color: cores.branco,
-   },
-});
